4.21

+++ part4/blogilista/controllers/blogs.js
@@ -38,8 +38,20 @@ blogsRouter.post('/', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndRemove(request.params.id)
-  response.status(204).end()
+  const blog = await Blog.findById(request.params.id)
+
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  console.log("decodedToken", decodedToken)
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: 'token invalid' })
+  }
+  const user = await User.findById(decodedToken.id)
+  console.log("blos.user", blog.user.toString())
+  console.log("user.toString", user._id.toString())
+  if (blog.user.toString() === user._id.toString()) {
+    await Blog.findByIdAndRemove(request.params.id)
+    response.status(204).end()
+  } else response.status(401).json({error: 'only the user who has posted the blog can delete it'})
 })
 
 blogsRouter.put('/:id', async (request, response) => {
